+++ src/main/scala/com/dataintuitive/viash/wrapper/BashWrapper.scala
@@ -39,6 +39,7 @@ object BashWrapper {
   def escapeViash(str: String, quote: Boolean = false, newline: Boolean = false): String = {
     Bash.escape(str, quote = quote, newline = newline)
       .replaceAll("\\\\\\$VIASH_DOLLAR\\\\\\$", "\\$")
+      .replaceAll("\\\\\\\\VIASH_SLASH\\\\\\\\", "\\\\")
       .replaceAll("\\\\\\$VIASH_", "\\$VIASH_")
       .replaceAll("\\\\\\$\\{VIASH_", "\\${VIASH_")
   }