+++ src/test/resources/testbash/tests/check_outputs.sh
@@ -35,14 +35,14 @@ grep -q 'head of resource1: |if you can read this,|' output.txt
 grep -q 'Parsed input arguments.' log.txt
 
 echo ">>> Checking whether output is correct with minimal parameters"
-./testbash "resource2.txt" --real_number 123.456 --whole_number=789 -s 'my$weird#string' \
+./testbash "resource2.txt" --real_number 123.456 --whole_number=789 -s "my\$weird#string\"\"\"" \
   > output2.txt
 
 [[ ! -f output2.txt ]] && echo "Output file could not be found!" && exit 1
 grep -q 'input: |resource2.txt|' output2.txt
 grep -q 'real_number: |123.456|' output2.txt
 grep -q 'whole_number: |789|' output2.txt
-grep -q 's: |my$weird#string|' output2.txt
+grep -q "s: |my\$weird#string\"\"\"|" output2.txt
 grep -q 'truth: |false|' output2.txt
 grep -q 'falsehood: |true|' output2.txt
 grep -q 'reality: ||' output2.txt